ISSUE-006 앱 패키지명 변경
