+++ modules/app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     package="com.yapp">
 
     <application
-        android:name=".App"
+        android:name="kr.merang.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication">
         <activity
-            android:name=".ui.MainActivity"
+            android:name="kr.co.yapp.merang.ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MyApplication">
