--- modules/app/src/main/java/com/yapp/App.kt
@@ -1,7 +1,0 @@
-package com.yapp
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class App : Application()
