+++ modules/app/src/main/java/kr/co/yapp/merang/App.kt
@@ -1,0 +1,7 @@
+package kr.co.yapp.merang
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
