--- modules/data/src/main/java/com/yapp/data/di/RepositoryModule.kt
@@ -1,18 +1,0 @@
-package com.yapp.data.di
-
-import com.yapp.data.repo.UserRepository
-import com.yapp.data.repo.UserRepositoryImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal abstract class RepositoryModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindUserRepository(repo: UserRepositoryImpl): UserRepository
-}
