--- modules/data/src/main/java/com/yapp/data/repo/UserRepository.kt
@@ -1,7 +1,0 @@
-package com.yapp.data.repo
-
-// TODO: UserRepository domain 모듈에 위치해야 함
-interface UserRepository {
-
-    fun getName(): String
-}
