--- modules/data/src/main/java/com/yapp/data/repo/UserRepositoryImpl.kt
@@ -1,10 +1,0 @@
-package com.yapp.data.repo
-
-import javax.inject.Inject
-
-internal class UserRepositoryImpl @Inject constructor() : UserRepository {
-
-    override fun getName(): String {
-        return "Android"
-    }
-}
