+++ modules/data/src/main/java/kr/co/yapp/merang/data/di/RepositoryModule.kt
@@ -1,0 +1,18 @@
+package kr.co.yapp.merang.data.di
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import kr.co.yapp.merang.data.repo.UserRepository
+import kr.co.yapp.merang.data.repo.UserRepositoryImpl
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal abstract class RepositoryModule {
+
+    @Binds
+    @Singleton
+    abstract fun bindUserRepository(repo: UserRepositoryImpl): UserRepository
+}
