+++ modules/data/src/main/java/kr/co/yapp/merang/data/repo/UserRepository.kt
@@ -1,0 +1,7 @@
+package kr.co.yapp.merang.data.repo
+
+// TODO: UserRepository domain 모듈에 위치해야 함
+interface UserRepository {
+
+    fun getName(): String
+}
