+++ modules/data/src/main/java/kr/co/yapp/merang/data/repo/UserRepositoryImpl.kt
@@ -1,0 +1,10 @@
+package kr.co.yapp.merang.data.repo
+
+import javax.inject.Inject
+
+internal class UserRepositoryImpl @Inject constructor() : UserRepository {
+
+    override fun getName(): String {
+        return "Android"
+    }
+}
