--- modules/presentation/src/main/java/com/yapp/ui/MainViewModel.kt
@@ -1,14 +1,0 @@
-package com.yapp.ui
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import javax.inject.Inject
-
-@HiltViewModel
-class MainViewModel @Inject constructor() : ViewModel() {
-
-    private val _name = MutableStateFlow("")
-    val name = _name.asStateFlow()
-}
